redo transcription waveform, kinda crashing but the transcription is kinda working!
