+++ src/core/apps/transcription/transcribe.rs
@@ -1,19 +1,17 @@
+use crate::apps::waveform::WaveformCanvas;
 use crate::engine::audio::{self, transcription::TranscriptionEngine};
 use crate::engine::{Application, EngineState};
 use crate::rasterizer::fill;
-use crate::rasterizer::fill_rect_buffer;
-use crate::ui::UiText;
 #[cfg(not(target_os = "ios"))]
 use dialoguer::Select;
 use std::io::{self, Write};
 
 const BG: (u8, u8, u8, u8) = (12, 14, 20, 255);
-/// Bars across the bottom for mic level (not Whisper decode).
-const WAVEFORM_COLS: usize = 160;
 
 pub struct TranscribeApp {
     listener: Option<audio::AudioListener>,
     engine: TranscriptionEngine,
+    wave: WaveformCanvas,
     last_caption_out: String,
 }
 
@@ -22,6 +20,7 @@ impl TranscribeApp {
         Self {
             listener: None,
             engine: TranscriptionEngine::new(),
+            wave: WaveformCanvas::new(),
             last_caption_out: String::new(),
         }
     }
@@ -34,58 +33,11 @@ impl TranscribeApp {
         println!("{caption}");
         let _ = io::stdout().flush();
     }
-
-    /// Recent mono samples → bottom strip of vertical bars (live mic sanity check).
-    fn draw_waveform_strip(
-        buffer: &mut [u8],
-        width: usize,
-        height: usize,
-        mono: &[f32],
-    ) {
-        let strip_h = ((height as f32) * 0.12).max(28.0).round() as i32;
-        let y0 = height as i32 - strip_h;
-        let strip_color = (22, 26, 34, 255);
-        for y in y0.max(0)..height as i32 {
-            fill_rect_buffer(buffer, width, height, 0, y, width as i32, y + 1, strip_color);
-        }
-        if mono.is_empty() || width == 0 {
-            return;
-        }
-        let cols = WAVEFORM_COLS.min(width);
-        let chunk = (mono.len() / cols).max(1);
-        let start = mono.len().saturating_sub(chunk * cols);
-        let bar_color = (80, 200, 220, 255);
-        let col_w = (width as f32 / cols as f32).max(1.0);
-        for c in 0..cols {
-            let i0 = start + c * chunk;
-            let i1 = (i0 + chunk).min(mono.len());
-            let peak = mono[i0..i1]
-                .iter()
-                .map(|s| s.abs())
-                .fold(0.0f32, f32::max);
-            let amp = (peak * 6.0).clamp(0.0, 1.0);
-            let bar_h = (amp * strip_h as f32).max(2.0).round() as i32;
-            let x0 = ((c as f32) * col_w).floor() as i32;
-            let x1 = (((c + 1) as f32) * col_w).ceil() as i32;
-            let bx1 = (x1).min(width as i32);
-            let by0 = height as i32 - bar_h;
-            fill_rect_buffer(
-                buffer,
-                width,
-                height,
-                x0.max(0),
-                by0.max(y0),
-                bx1.max(x0 + 1),
-                height as i32,
-                bar_color,
-            );
-        }
-    }
 }
 
 impl Application for TranscribeApp {
     fn setup(&mut self, _state: &mut EngineState) -> Result<(), String> {
-        println!("transcribe: window + live waveform strip · transcript on stdout · Esc to quit");
+        println!("transcribe: waveform in window · transcript on stdout · Esc to quit");
         let _ = io::stdout().flush();
 
         let all_devices = audio::devices();
@@ -166,56 +118,23 @@ impl Application for TranscribeApp {
     }
 
     fn tick(&mut self, state: &mut EngineState) {
-        fill(&mut state.frame, (BG.0, BG.1, BG.2, BG.3));
-
-        let shape = state.frame.shape();
-        let height = shape[0] as usize;
-        let width = shape[1] as usize;
-        if width == 0 || height == 0 {
+        if self.listener.is_none() {
+            fill(&mut state.frame, (BG.0, BG.1, BG.2, BG.3));
             return;
         }
 
-        let font_px = 18.0_f32 * state.f3_ui_scale_multiplier();
-
-        let wave_mono = if let Some(listener) = &self.listener {
-            let channels = listener.get_samples_by_channel();
-            let sr = listener.buffer().sample_rate();
-            self.engine.process_snapshot(sr, &channels);
-            channels.into_iter().next().unwrap_or_default()
-        } else {
-            Vec::new()
+        let (channels, sr) = {
+            let l = self.listener.as_ref().expect("checked above");
+            (l.get_samples_by_channel(), l.buffer().sample_rate())
         };
 
-        let caption = if self.listener.is_some() {
-            self.engine.caption().to_string()
-        } else {
-            "No audio listener.".to_string()
-        };
+        self.engine.process_snapshot(sr, &channels);
 
+        let caption = self.engine.caption().to_string();
         self.log_caption_to_stdout(&caption);
 
-        let buf = state.frame_buffer_mut();
-        Self::draw_waveform_strip(buf, width, height, &wave_mono);
-
-        let display = if self.listener.is_some() {
-            format!("{}\n\n{}", self.engine.device_hint(), caption)
-        } else {
-            caption.clone()
-        };
-        let ui = UiText {
-            text: display,
-            x1_norm: 0.04,
-            y1_norm: 0.04,
-            x2_norm: 0.96,
-            y2_norm: 0.82,
-            color: (235, 238, 245, 255),
-            hitboxes: false,
-            baselines: false,
-            font_size_px: font_px.max(10.0),
-        };
-        if let Err(e) = ui.render(buf, width, height) {
-            eprintln!("transcribe: UiText render error: {e}");
-        }
+        let l = self.listener.as_ref().expect("checked above");
+        self.wave.tick_draw(state, l);
     }
 
     fn on_key_char(&mut self, _state: &mut EngineState, ch: char) {
